Treat layers uniformly, including last (dev) layer
The main function is just a placeholder anyway but for now, a few more
checks so we avoid exceptions and get a printout when there is no config
output from last layer.

The more important change is that we now treat all the layer uniformly.
The last layer is that of device configuration and thus far I have
imagined we would keep the device config as separate trees, one per
device, since that's sort of the natural way and shape of it.. however,
that means we must make exceptions in the transaction engine. After some
more thought and discussion, it was decided against this, so now we will
join the device config under one single tree so that it looks uniform to
the transaction engine, yay!