Hack XR model shutdown leaf into empty leaf
I have no idea why Cisco has chosen to use a presence container for
modeling shutdown on an interface. It's really weird.

Interfaces on Cisco IOS XR routers are shutdown per default and we want
to remove that. In the long term, this will happen automatically as we
will effectively replace the entire config (although via diff
computation). We don't do that yet though, so need to find a way to
remove just the shutdown.

I actually presumed it was a leaf of type empty (it used to be in the
older non-UM XR YANG models) and implemented supported in the YANG model
for treating type empty as a boolean and when set to False, that would
map to removing the leaf. That works, but obviously not for
P-containers, since those wouldn't be mapped to bools in Acton.

However, over NETCONF, the XML serialization of creating an empty leaf
or creating a P-container are the same. And removing them is also the
same. So by just hacking the YANG model to claim `shutdown` is now a
type empty leaf, we can work with in Acton and the XR device is
none-the-wiser.