Read configs from disk via argv
In addition to accepting configs via the HTTP server, we now bring back
the old behavior of reading files from disk that are passed in as an
argument. This is the best of both worlds, since you don't have to pass
in a config unless you want to. Passing in a file is very simple for
testing though, one less thing to do!