Stop rewriting Junos unions
We can now handle union types with defaults. It is a bit more work in
the code to work with mixed union types (like int and str) because they
are right now stored as Acton value type, but that will change once
Acton supports union types. On the other hand, we can now use a
"vanilla" model from Juniper.