Uese the location street address in l3pvn-svc
Using the street address under the l3vpn asvc location is a more
realistic example of how the model is (supposed to be) used.

The mapping from city name to a router name & interface happens through
a silmple dictionary lookup for now, this can be expanded on later to
move to a proper asset database.