Await device support
The Orchestron transaction engine & device manager offers two callbacks,
the first is when the transaction is done but before the configuration
changes has been pushed to the devices and the second is when the device
push is complete. These are called done and complete respectively and
are the second and third argument to

    Session.edit_config(diff, done, complete)

The application of configuration from files fed in as arguments now uses
the await-device option so that we await the device push to be complete
before proceeding with the next configuration file. This radically
changes the flow, where previously we had a simple sleep of 3 seconds
between files but no actual wait on anything, so if devices were not yet
ready we would still just apply config files. Now the first config file
will wait for the device to become ready before proceeding with the next
file.

For the RESTCONF API, the default remains asynchronous but by setting
the header await-device=true the request will await push to device is
complete.