Decouple ping (network) test from get-config
Right now we get the device running configuration via NETCONF directly
from the device. This may take some time, first to pull the notconf
container (for netconf-console2) and also to fetch the configs. We now
only limit the duration of `test-ping` to two minutes.

A more proper solution would be to use Orchestron itself to get config
which we can then save to disk via curl, but as of right now, we can't
fully parse the <get-config> response for all supported device types ...