Remove mandatory constraints from top-level
If a mandatory constraint is used in a leaf (in a non-presence container
even), then it becomes a top-level constraint. sysrepo enforces these
when installing YANG modules. We could provide the data with the
"--init-data" option, but seeing as the real cRPD also lies here and
doesn't provide a value, we just remove the mandatory constraint.