Fetch conf and add /diff to HTTP API
It is now possible to get a diff of the running config on the device and
our target config. This should make it easier to bring our transforms in
line, for example adding configuration that is currently on the devices
so that we later can turn on the diff reconciliation step.

Upgraded Orchestron to get the config-fetching during commit dance.