Add a bunch of default / required config for SRL
Nokia SR Linux has a bunch of default config, likea  mgmt network
instance. It's not really possible to remove this because there is
other (default? and hidden??) config that references. Thus, we must
configure this configuration or else our future diff-based config
reconciliation dance will attempt to remove it, in order to bring the
device in line with the Orchestron/SORESPO target config.