Do diff-based config push
Upgrade Orchestron dep, which brings in new functionality that computes
diff between currently running config on device and the target config,
which is then sent to the device. With this, we will remove
configuration from the device in order to bring it in line with the
target configuration in Orchestron / SORESPO, effectively overwriting
the configuration on the device.