Flip to await device completion etc by default
Our default behavior has been to act asynchronously but have exposed the
Await-Device header to optionally await that configuration has been
pushed to the device before the HTTP API returns a response.

This flips the default so that we now await device completion per
default. If a user wants async, they can request that by sending the
header: Async: true

I think this better reflects naive users intuition. Most systems are
synchronous and it's quite nice from a control flow perspective that
when the Orchestron HTTP API returns, the configuration has been
entirely applied, so if you are doing this from a simple script, the
script can block on the HTTP API call and once it is done, it can
proceed to the next thing. I think it will be even better when we
integrate the concepts of transform health so that we can also ensure
that a transform is healthy after a configuration change or at least
didn't deteriorate (broken before = broken after is acceptable). For
anyone that has a concurrent system or is used to dealing with async, it
is trivial to enable.

The make targets have been renamed. We have -async or -wait explicitly
in the name, which I think is OK, since even if you use the defauly
synchronous, having -wait as the suffix hints that there are other
options there.