Update golden test values
With the gen3 parsers in use for all gdata conversion, the ordering may
be a bit different (schema order). Also we no longer emit empty
leaf-lists and parse union types correctly!

We also use the deterministic=True flag for all gdata printers because
in a multi-layer TTT system the order of merge operations for transforms
is non-deterministic. Semantically the output is stable, but the order
of some nodes where the parent nodes overlap between different
transforms may change.